organize example page and functions
